+++ gamepatches.py
@@ -192,10 +192,10 @@ DUNGEON_FINISH_EXIT_SCEN = {
     # stage, room, index
     SV: ("B100_1", 0, 1),
     ET: ("B210", 0, 0),
-    LMF: ("F300_4", 0, 3),
-    AC: ("B101_1", 0, 3),
-    SSH: ("B301", 0, 4),
-    FS: ("B201_1", 0, 2),
+    LMF: ("F300_5", 0, 0),
+    AC: ("B101_1", 0, 2),
+    SSH: ("B301", 0, 3),
+    FS: ("B201_1", 0, 1),
     SK: ("F407", 0, 1),
 }
 
@@ -2375,7 +2375,7 @@ class GamePatcher:
             rng.shuffle(locs)
             # print(locs)
 
-            for ((id, room), (params, actor_name)) in zip(
+            for (id, room), (params, actor_name) in zip(
                 locs,
                 params,
             ):
@@ -2819,7 +2819,7 @@ class GamePatcher:
             start_flags_write.write(struct.pack(">H", flag))
         start_flags_write.write(bytes.fromhex("FFFF"))
         # itemflags
-        for (flag, count) in self.startitemflags.items():
+        for flag, count in self.startitemflags.items():
             assert flag < 0x1FF
             assert count < 0x7F
             start_flags_write.write(struct.pack(">H", (count << 9) | flag))
@@ -2845,6 +2845,12 @@ class GamePatcher:
         start_flags_write.write(struct.pack(">H", 0))
         # Start health.
         start_flags_write.write(struct.pack(">B", self.starting_full_hearts))
+        # start interface choice
+        interface_choice_num = ["Standard", "Light", "Pro"].index(
+            self.placement_file.options["interface"]
+        )
+        start_flags_write.write(struct.pack(">B", interface_choice_num))
+
         startflag_byte_count = len(start_flags_write.getbuffer())
         if startflag_byte_count > 512:
             raise Exception(